* Weather folder support * Weather.ini files translation goes directly after weather plugin translation in full langpack file * Subfolders created automatically for /out parameter * Untranslated folder created automatically (one-level only, in other words, with /untranslated:"/path/to/untranslated", folder /path/to have to be existed, "untranslated" will be created inside /path/to)
ATTENTION: now for /out parameter you need specify a upper-level folder, not "plugins" as before. In other words, with /out:"path/to/folder" a folder with name "folder" will be created in /path/to as well as  subfolders /path/to/folder/Plugins and /path/to/folder/Weather will be created too.

Careful testing needed :)

git-svn-id: http://svn.miranda-ng.org/miranda-ng/trunk@5291 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c