IEHistory: partially revert of r10283 : either do it right, or don't do it at all. capitalization was chosen wisely because it's already used by Miranda and thus doesn't bloat the language files. I also chose to use the very same naming as History++ did.
It's our responsibility as developers to not bloat language files. They are already bloated enough, don't extend it even further. (and I wish someone would write a "string search program" to find similar already translated/used strings :P)

You may decapitialize it again, but then also do it for History++ and Import. (we could actually different dialog name, as Import uses the same spelling, but as said, History++ doesn't. The menu what so ever must be upper case. Or change it in History++ as well)

git-svn-id: http://svn.miranda-ng.org/miranda-ng/trunk@10305 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c